Change Instance Ownership
The idea here is the allow the instance to be relocated. It can receive
a new user and a new project.

This code allows the changes for a new user and a new project, the user
change needs more verification but the project already has quotas check
and it moves the data of the instance from one project to another.

Change-Id: I11cea028f4ad153bfb88efc45cd762fcae15ae79